Move task sources into one place
