+++ crates/task/src/lib.rs
@@ -1,10 +1,8 @@
 //! Baseline interface of Tasks in Zed: all tasks in Zed are intended to use those for implementing their own logic.
 #![deny(missing_docs)]
 
+pub mod oneshot_source;
 pub mod static_source;
-mod static_task;
-
-pub use static_task::StaticTask;
 
 use collections::HashMap;
 use gpui::ModelContext;
